fix: enable selector, move EnableVSP button to top
