+++ app/components/views/TicketsPage/PurchaseTab/PurchaseTab.module.css
@@ -88,7 +88,7 @@
 }
 
 .enableVSPButton {
-  margin-top: 10px;
+  margin-bottom: 40px;
 }
 
 .enableVSPWrapper {
